Specialized clipboard service
