+++ src/sql/platform/clipboard/electron-browser/clipboardService.ts
@@ -5,70 +5,24 @@
 
 import { IClipboardService } from 'sql/platform/clipboard/common/clipboardService';
 import { IClipboardService as vsIClipboardService } from 'vs/platform/clipboard/common/clipboardService';
+import { ClipboardService as BrowserClipboardService } from 'sql/platform/clipboard/browser/clipboardService';
 import { clipboard, nativeImage } from 'electron';
-import { URI } from 'vs/base/common/uri';
 
-export class ClipboardService implements IClipboardService {
+export class ClipboardService extends BrowserClipboardService {
 	_serviceBrand: any;
 
 	constructor(
-		@vsIClipboardService private _vsClipboardService: vsIClipboardService
-	) { }
+		@vsIClipboardService _vsClipboardService: vsIClipboardService
+	) {
+		super(_vsClipboardService);
+	}
 
 	/**
 	 * Writes the input image as a dataurl to the clipbaord
 	 */
-	writeImageDataUrl(data: string): void {
+	async writeImageDataUrl(data: string): Promise<boolean> {
 		let image = nativeImage.createFromDataURL(data);
 		clipboard.writeImage(image);
-	}
-
-	writeText(text: string): Promise<void> {
-		return this._vsClipboardService.writeText(text);
-	}
-
-	/**
-	 * Reads the content of the clipboard in plain text
-	 */
-	readText(): Promise<string> {
-		return this._vsClipboardService.readText();
-	}
-	/**
-	 * Reads text from the system find pasteboard.
-	 */
-	readFindText(): string {
-		return this._vsClipboardService.readFindText();
-	}
-
-	/**
-	 * Writes text to the system find pasteboard.
-	 */
-	writeFindText(text: string): void {
-		this._vsClipboardService.writeFindText(text);
-	}
-
-	/**
-	 * Writes resources to the system clipboard.
-	 */
-	writeResources(resources: URI[]): void {
-		this._vsClipboardService.writeResources(resources);
-	}
-
-	/**
-	 * Reads resources from the system clipboard.
-	 */
-	readResources(): URI[] {
-		return this._vsClipboardService.readResources();
-	}
-
-	/**
-	 * Find out if resources are copied to the clipboard.
-	 */
-	hasResources(): boolean {
-		return this._vsClipboardService.hasResources();
-	}
-
-	readTextSync(): string | undefined {
-		return this._vsClipboardService.readTextSync();
+		return true;
 	}
 }
